Refactor WashingtonPostCollection class to have inner classes for Document
